# aigis/Models/RegisterRequest.cs
namespace aigis.Models
{
    public class RegisterRequest
    {
        public string Nombre { get; set; }
        public string Correo { get; set; }
        public string Contrasena { get; set; }
        public string Direccion { get; set; }
        public string Telefono { get; set; }
    }
}