# aigis/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "aigis, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/AyudaUsuarios": {
      "post": {
        "tags": [
          "AyudaUsuarios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AyudaUsuarios"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/AyudaUsuarios/{id}": {
      "get": {
        "tags": [
          "AyudaUsuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "AyudaUsuarios": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "titulo": {
          "type": "string"
        },
        "problema": {
          "type": "string"
        },
        "usuarioId": {
          "type": "string"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "LoginRequest": {
      "type": "object",
      "properties": {
        "correo": {
          "type": "string"
        },
        "contrasena": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "RegisterRequest": {
      "type": "object",
      "properties": {
        "nombre": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "contrasena": {
          "type": "string"
        },
        "direccion": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  }
}